refactor: move composite action logic into own file
